Split currency overlay supportability
