+++ docs/architecture/CODEBASE-REVIEW-LEDGER.md
@@ -8006,3 +8006,25 @@ This ledger records cleanup and structural review evidence for RFC-0036.
   supportability functions until those are split.
 - Wiki decision: no wiki source change required; this is internal test modularity cleanup with no
   route, payload, supported-feature, or operator-contract change.
+
+## BACKEND-REVIEW-20260601-326: Currency overlay supportability module split
+
+- Date: 2026-06-01
+- Scope: `src/api/services/construction_method_supportability.py`,
+  `src/api/services/construction_currency_overlay_supportability.py`, construction method
+  supportability tests, and this ledger.
+- Finding: after the liquidity split, construction method supportability still mixed
+  currency-overlay policy derivation, missing-pair detection, and status handling with the
+  regime-stress facade.
+- Action: extracted currency-overlay supportability into a focused helper module while preserving
+  the existing method-supportability facade imports for callers.
+- Status: hardened
+- Evidence: focused Ruff and format checks passed for the touched source/test files; focused mypy
+  passed for the method and currency-overlay supportability source modules; focused method
+  supportability regressions passed with 2 tests; OpenAPI quality gate passed; API vocabulary
+  inventory validate-only gate passed; `git diff --check` passed; service leakage scan found no
+  router/HTTP imports in service modules.
+- Follow-up: move direct currency-overlay tests into a dedicated test module, then split
+  regime-stress supportability out of the method facade.
+- Wiki decision: no wiki source change required; this is internal service modularity cleanup with no
+  route, payload, supported-feature, or operator-contract change.
